[Refactor] Member Role -> Enum값으로 변경 #25
