+++ src/main/java/com/appcenter/BJJ/domain/member/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.appcenter.BJJ.domain.member.repository;
+package com.appcenter.BJJ.domain.member;
 
 import com.appcenter.BJJ.domain.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
