+++ src/main/java/com/appcenter/BJJ/domain/member/MemberRole.java
@@ -1,0 +1,8 @@
+package com.appcenter.BJJ.domain.member;
+
+import lombok.Getter;
+
+@Getter
+public enum MemberRole {
+    GUEST, USER
+}
